Moved actions folder to root of the project
