+++ app/api/conversations/[conversationId]/seen/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 
-import getCurrentUser from "@/app/actions/getCurrentUser";
+import getCurrentUser from "@/actions/getCurrentUser";
 import { pusherServer } from "@/app/libs/pusher";
 import prisma from "@/app/libs/prismadb";
