+++ app/api/settings/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import getCurrentUser from "@/app/actions/getCurrentUser";
+import getCurrentUser from "@/actions/getCurrentUser";
 import prisma from "@/app/libs/prismadb";
 
 /**
